Another round of code warning fixed. This includes - commenting out unused function parameters in C++ (it's allowed in C++) - move static function prototypes from shared headers to the files where they are actually defined. - fixing the issues with plplotP.h needed config.h - more casting fixes and other minor tweaks - extra braces for clarity in nested if / else statements to avoid ambiguity.
svn path=/trunk/; revision=11994

+++ drivers/xwin.c
@@ -1102,8 +1102,8 @@ Init( PLStream *pls )
     (void) XGetGeometry( xwd->display, dev->window, &root, &x, &y,
         &dev->width, &dev->height, &dev->border, &xwd->depth );
 
-    dev->init_width  = dev->width;
-    dev->init_height = dev->height;
+    dev->init_width  = (long) dev->width;
+    dev->init_height = (long) dev->height;
 
 // Set up flags that determine what we are writing to
 // If nopixmap is set, ignore db
